[PEA] At loop header, only materialize objects that are not live
Add a HashMapBench benchmark, which with default parameters decreases
allocation rate by 25% with PEA turned on. We can avoid allocation of
the value object in HashMap::replace if the key is not present. Before
this patch, PEA could not optimize it because there is a loop in the
inline tree, causing escaping of all tracked objects. With this patch,
PEA finds that the loop has nothing to do with the value object and does
not materialize it.