[PEA] Mark get_field'ed virtual objects as aliases
Prior to this patch, we replaced the get_field'ed object with the
virtual object saved in the state. The problem is that if we are to
support loops, it would be hard to undo this change when merging
backedges.

In this patch, we mark the following nodes as aliases:
* LoadP - get_field'ed objects with -XX:-UseCompressedOops
* DecodeN - get_field'ed objects with compressed oops
* CastPP - null-checked objects on DecodeN's because they lack
non-null typing data, null checks are inserted for calls and loads
* CheckCastPP - instanceof checks on CastPP's because they lack exact
  typing data

Now at any point of execution, there can be many nodes aliasing to the
same allocation. When we materialize an object, we need to make sure we
replace all nodes that alias to that object with the materialized object.