WIP. Only emit one storestore for an object allocation.
Draft version where we check:
1) the current method has the same receiver as the caller
2) both the current and parent method are initializers

These points would indicate that the callee is a superclass constructor
and does not need a trailing barrier.

This case does not work in the case where the parent ctor writes a final
and the child ctor does not. The parent ctor will see its a superclass
ctor and not emit a barrier. The child ctor will see it did not write a
final and not emit a barrier. We need to transfer `write_final` and
other `write_*` fields from child Parse to parent Parse(WIP). Its not
that trivial since code is written such that Parse instances are
independant of one another.

Turns out we pass the testcases anyway. A StoreStore is emitted when
expanding the IntializeNode. We could do better testing by testing
against an earlier phase.

I found an easier way to do this which I am submitting a PR for. But
saving this in case that approach gets reject and we need to revert to
this one.